+++ packages/sdk/src/wallet/substrates/HTTPWalletJSON.ts
@@ -38,7 +38,7 @@ import {
 import { WERR_REVIEW_ACTIONS } from '../WERR_REVIEW_ACTIONS.js'
 import { WERR_INVALID_PARAMETER } from '../WERR_INVALID_PARAMETER.js'
 import { toOriginHeader } from './utils/toOriginHeader.js'
-import { normalizeWalletJsonTx, walletJsonReplacer } from './utils/jsonByteEncoding.js'
+import { normalizeBRC100WalletByteFields, stringifyBRC100 } from '../BRC100ByteEncoding.js'
 import WERR_INSUFFICIENT_FUNDS from '../WERR_INSUFFICIENT_FUNDS.js'
 
 function deserializeWalletError(data: any): Error | undefined {
@@ -102,10 +102,10 @@ export default class HTTPWalletJSON implements WalletInterface {
           'Content-Type': 'application/json',
           ...(origin ? { Origin: origin, Originator: origin } : {})
         },
-        body: JSON.stringify(args, walletJsonReplacer)
+        body: stringifyBRC100(args)
       })
 
-      const data = normalizeWalletJsonTx(await res.json())
+      const data = normalizeBRC100WalletByteFields(await res.json())
 
       // Check the HTTP status on the original response
       if (!res.ok) {
@@ -118,7 +118,7 @@ export default class HTTPWalletJSON implements WalletInterface {
           args,
           message: data.message ?? `HTTP Client error ${res.status}`
         }
-        throw new Error(JSON.stringify(err))
+        throw new Error(stringifyBRC100(err))
       }
       return data
     }